% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bakery5000}
\alias{bakery5000}
\title{Items of 5000 transactions in a chain of bakery stores.}
\format{
A data frame with 5000 rows (transactions) and 50 variables (items):
\describe{
  \item{cakes}{F1--F10 belong to cakes}
  \item{tarts}{F11--F20 belong to tarts}
  \item{cookies}{F21--F30 belong to coolies}
  \item{pastries}{F31--40 belong to pastries}
  \item{drinks}{D1--D10 belong to drinks}
}
}
\source{
\url{http://users.csc.calpoly.edu/~dekhtyar/466-Spring2018/labs/lab01.html}
}
\usage{
bakery5000
}
\description{
A sub-sample of the Extended Bakery Dataset. The full dataset provides the transaction
details in a chain of bakery stores. 
'F' in the colnames represents Food and D is for Drink.
}
\examples{
data(bakery5000)
}
\keyword{datasets}
