% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahub.R
\name{data.hub}
\alias{data.hub}
\title{Data Analysis using penalized hub model}
\usage{
data.hub(data, M, Lam, bootstrap = 0)
}
\arguments{
\item{data}{observed group data, the first \eqn{M} colunmns are corresponding to hub nodes}

\item{M}{size of potential hub set}

\item{Lam}{a vector of tunning parameter \eqn{\lambda}}

\item{bootstrap}{number of bootstrap times}
}
\value{
a matrix of hub nodes weight for different \code{Lam}
}
\description{
Data Analysis using penalized hub model
}
\examples{
data(bakery5000)
Lam = seq(0.015,0.035,0.005)
data.hub(bakery5000,40,Lam,0) # full data
data.hub(bakery5000,40,Lam,2) # bootstrapped data
  
}
