% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMM.R
\name{EMM.hub}
\alias{EMM.hub}
\title{Multiple initial points (rho) version of function \code{\link{EM.hub}} (Deprecated now)}
\usage{
EMM.hub(G, M, lam, iter = 1000, tol = 0.01)
}
\arguments{
\item{G}{observed group data, a T*n matrix}

\item{M}{initial user-specifed number of components}

\item{lam}{tuning parameter for component selection, lambda*Df; if lam=0, becomes a normal EM without penalty}

\item{iter}{maximum iteration steps}

\item{tol}{tolerated chnage for log-likelihood}
}
\value{
a list of components
\item{A}{M*n matrix containg estimated correlation among nodes}
\item{rho}{M-dimensional vector containing estimated component weight}
\item{l}{loglikelihood}
\item{iteration}{number of iteration steps used to converge}
}
\description{
Multiple initial points (rho) version of function \code{\link{EM.hub}} (Deprecated now)
}
\examples{
n0 = 5; n=100
A0 = GenA(100,n0,0.4,0.1,rep(0.05,n))
G0 = GenG(A0,1000,c(0.2,rep(0.8/n0,n0)))
M = 10; EMM.hub(G0,M,0.04)
M = 20; EMM.hub(G0,M,0.035)
\donttest{M = 50; EMM.hub(G0,M,0.019)}

}
