% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generation.R
\name{GenA}
\alias{GenA}
\title{Generate a simple pattern adjacent matrix \eqn{A}}
\usage{
GenA(n, n0, pf, pr, pi)
}
\arguments{
\item{n}{number of nodes}

\item{n0}{number of hub nodes}

\item{pf}{\eqn{A_{ij}} if node j in the follower set of hub i}

\item{pr}{\eqn{A_{ij}} if node j does not belong to the follower set of hub i}

\item{pi}{a n-dimensional vector containing probability of appearance in non-hub group}
}
\value{
a (n0+1)*n matrix, the first row is \code{pi}
}
\description{
Generate a simple pattern adjacent matrix \eqn{A}
}
\examples{
pi = rep(0.05,100)
A0 = GenA(100,5,0.4,0.1,pi)

}
