% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generation.R
\name{GenG}
\alias{GenG}
\title{Generate grouped data \eqn{G}}
\usage{
GenG(A, T, rho)
}
\arguments{
\item{A}{(n0+1)*n-dimensional adjacent matrix}

\item{T}{number of observations (groups)}

\item{rho}{a (n0+1) dimensional vector containing hub weight}
}
\value{
T*n matrix
}
\description{
Generate grouped data \eqn{G}
}
\examples{
n = 100; n0 = 5; T = 1000
A0 = GenA(n,n0,0.4,0.1,rep(0.05,100))
rho = c(0.2,rep(0.8/n0,n0))
G0 = GenG(A0,T,rho)

}
