% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{passerines}
\alias{passerines}
\title{Grouping behavior of passerines}
\format{
A data frame with 109 rows and 63 variables:
}
\source{
\url{https://zookeys.pensoft.net/articles.php?id=6934}
}
\usage{
passerines
}
\description{
The dataset includes 63 color-marked passerines in Australia for daily observations
}
\examples{
data(passerines)
}
\keyword{datasets}
