% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{phub}
\alias{phub}
\title{Modified EM algorithm for hub model component selection}
\usage{
phub(G, A, rho, lam, pen.type = c("plog", "plasso", "log"),
  iter.max = 1000, tol = 1e-04)
}
\arguments{
\item{G}{observed group data, a T*n matrix}

\item{A}{a matrix containg correlation among nodes; first row is the nonleader case}

\item{rho}{a vector composed of component weight; prior probability of components.}

\item{lam}{tuning parameter for component selection, degenrate to a standard EM without penalty if lam=0}

\item{pen.type}{type of penalty, including 'log': penalization for logarithm of all components;
'plog': penalization for logarithm of partial components, rho0 (nonhub component) is always kept;
'plasso': penalization for lasso form of partial compents}

\item{iter.max}{maximum iteration steps}

\item{tol}{threshold for shrinking rho to 0}
}
\value{
a list of components
\item{A}{a matrix containg estimated correlation among nodes}
\item{rho}{a vector containing estimated component weight}
\item{l}{log-likelihood}
\item{iteration}{number of iterations used to converge}
}
\description{
Modified EM algorithm for hub model component selection
}
\examples{
n0 = 5; n=100
A0 = GenA(100,n0,0.4,0.1,rep(0.05,n))
G0 = GenG(A0,1000,c(0.2,rep(0.8/n0,n0)))
M = 10
A = matrix(runif((M+1)*n),nrow=(M+1)); diag(A[-1,])=1
rho = runif(M+1); rho = rho/sum(rho)
phub(G0,A,rho,0.04,pen.type="log")$rho
phub(G0,A,rho,0.04,pen.type="plog")$rho
phub(G0,A,rho,0.08,pen.type="plasso")$rho
}
