% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{EM.hub}
\alias{EM.hub}
\title{Modified EM algorithm for hub model component selection}
\usage{
EM.hub(G, A, rho, lam, iter = 1000, tol = 0.01)
}
\arguments{
\item{G}{observed group data, a T*n matrix}

\item{A}{a matrix containg correlation among nodes; first row is the nonleader case}

\item{rho}{a vector composed of component weight; prior probability of components.}

\item{lam}{tuning parameter for component selection, lambda*Df; if lam=0, becomes a normal EM without penalty}

\item{iter}{maximum iteration steps}

\item{tol}{tolerated chnage for log-likelihood}
}
\value{
a list of components
\item{A}{a matrix containg estimated correlation among nodes}
\item{rho}{a vector containing estimated component weight}
\item{l}{log-likelihood}
\item{iteration}{number of iterations used to converge}
}
\description{
Modified EM algorithm for hub model component selection
}
\examples{
n0 = 5; n=100
A0 = GenA(100,n0,0.4,0.1,rep(0.05,n))
G0 = GenG(A0,1000,c(0.2,rep(0.8/n0,n0)))
M = 11
A = matrix(runif(M*n),nrow=M); diag(A[-1,])=1
rho = runif(M); rho = rho/sum(rho)
EM.hub(G0,A,rho,0.03)

}
