% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{EM.hub}
\alias{EM.hub}
\title{Modified EM algorithm for hub model component selection}
\usage{
EM.hub(G, A, rho, lam, iter = 1000)
}
\arguments{
\item{G}{Observed group data, a T*n matrix}

\item{A}{M*n matrix containg correlation among nodes; first row is the nonleader case}

\item{rho}{M-dimensional vector composed of component weight; prior probability of components.}

\item{lam}{tunning parameter for component selection, lambda*Df}

\item{iter}{maximum iteration steps}
}
\value{
a list of components
\item{A}{M*n matrix containg estimated correlation among nodes}
\item{rho}{M-dimensional vector containing estimated component weight}
\item{l}{loglikelihood}
\item{iteration}{number of iteration steps used to converge}
}
\description{
Modified EM algorithm for hub model component selection
}
\examples{
n0 = 5; n=100
A0 = GenA(100,n0,0.4,0.1,rep(0.05,n))
G0 = GenG(A0,1000,c(0.2,rep(0.8/n0,n0)))
M = 11
A = matrix(runif(M*n),nrow=M); diag(A[-1,])=1
rho = runif(M); rho = rho/sum(rho)
pi = colMeans(G0)
EM.hub(G0,A,rho,0.03)

}
