% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{GenG}
\alias{GenG}
\title{Simulate grouped data}
\usage{
GenG(A, T, rho)
}
\arguments{
\item{A}{(n0+1)*n-dimensional adjacent matrix}

\item{T}{number of observations}

\item{rho}{a (n0+1) dimensional vector containing component(hub) weight}
}
\value{
\code{T}*n matrix
}
\description{
Simulate grouped data
}
\examples{
n0 = 5
A0 = GenA(100,n0,0.4,0.1,rep(0.05,100))
rho = c(0.2,rep(0.8/n0,n0))
G0 = GenG(A0,1000,rho)

}
\seealso{
\code{\link{GenA}}
}
