% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{phub}
\alias{phub}
\title{Modified EM algorithm for hub set selection in hub models with the null component}
\usage{
phub(
  G,
  A,
  rho,
  lam,
  pen.type = c("plog", "plasso", "log"),
  iter.max = 1000,
  tol = 1e-04
)
}
\arguments{
\item{G}{observed group data}

\item{A}{initial estimate of adjacent matrix, the first row contains the probabilities in non-hub group}

\item{rho}{a vector of hub weight}

\item{lam}{tuning parameter for component selection, degenrate to a standard EM without penalty if lam=0}

\item{pen.type}{type of penalty, including 'log': penalization for logarithm of all components;
'plog': penalization for logarithm of partial components except the null component;
'plasso': penalization for lasso form of partial components}

\item{iter.max}{maximum iteration steps}

\item{tol}{threshold for shrinking rho to 0}
}
\value{
a list of components
\item{A}{a matrix containg estimated correlation among nodes}
\item{rho}{a vector containing estimated component weight}
\item{l}{log-likelihood}
\item{iteration}{number of iterations used to converge}
}
\description{
Modified EM algorithm for hub set selection in hub models with the null component
}
\examples{
set.seed(2020)
n0 = 5; n=100; T=1000
A0 = GenA(n,n0,0.4,0.1,rep(0.05,n))
G0 = GenG(A0,T,c(0.2,rep(0.8/n0,n0)))
M = 10 
A = matrix(runif((M+1)*n),nrow=(M+1)); diag(A[-1,]) = 1
rho = runif(M+1); rho = rho/sum(rho)
phub(G0,A,rho,0.030,pen.type="log")$rho
phub(G0,A,rho,0.030,pen.type="plog")$rho
phub(G0,A,rho,0.8,pen.type="plasso")$rho

}
